fix: always save js module as .jsx file (#32)
some tools (e.g. `tsx`) will not parse jsx syntax in `.js` files, but
require an explicit `.jsx` extension. this is relevant when using the
generated data in scripts and running them with `tsx`.

this changes the default file extension for `JavaScriptImport`s to
`.jsx`.

in the future we may want to switch to `.tsx` since this will have to be
processed by a transpiler/bundler anyway.